artist: update from master and extract interfaces
