+++ package.json
@@ -10,7 +10,7 @@
     "test": "mocha -r ts-node/register 'test/**/*.ts'",
     "test:coverage": "nyc mocha -r ts-node/register 'test/**/*.ts'",
     "get:coverage": "nyc report --reporter=text-lcov",
-    "test:watch": "mocha ts-node/register 'test/**/*.ts' --watch",
+    "test:watch": "mocha --watch -r ts-node/register 'test/**/*.ts' ",
     "build": "npm test && tsc",
     "docs": " typedoc --options typedoc.json --name \"P07 - Digitalizando la coleccion de musica de los abuelos\" --includeVersion"
   },
